requirements.txt
Flask>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Mail>=0.10.0
Flask-Migrate>=4.1.0
Flask-SQLAlchemy>=3.1.1
WTForms>=3.1.2
psycopg2-binary>=2.9.9
cryptography>=42.0.0
python-dotenv>=1.0.1
openpyxl>=3.1.3
requests>=2.31.0
Pillow>=10.2.0
gunicorn>=21.2.0
email-validator>=2.1.0
# WeasyPrint  # opcional - para exportação PDF, instale separadamente
